README.md
The Qiskit Global Summer School is a two-week intensive summer program designed to empower the quantum researchers and developers of tomorrow with the know-how to explore the world of quantum computing, as well as refresh and sharpen the industry professional’s skills. This fifth-annual summer school will prepare participants for the imminent era of quantum utility. Using hands on examples, the syllabi will guide students through areas of near-term practical importance.